RF | move methods from class AddDescriptionRDVTest in the class _FreestyleTest
